requirements.txt
asgiref==3.8.1
autopep8==2.3.0
certifi==2024.6.2
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
coverage==7.5.4
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
django-seed==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==25.9.1
idna==3.7
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
openapi-codec==1.3.2
packaging==24.1
psycopg==3.1.19
psycopg2-binary==2.9.9
pycodestyle==2.12.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.1.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
